+++ src/text/str/nonul.rs
@@ -361,11 +361,10 @@ impl<const CAP: usize> StringNonul<CAP> {
     #[rustfmt::skip]
     pub const fn from_char(c: char) -> Result<Self, MismatchedCapacity> {
         let mut new = unwrap![ok? Self::new()];
-
-        if c as u32 != 0 {
+        if c != '\0' {
             let bytes = Char::to_utf8_bytes(c);
-            let len = Char::utf8_4bytes_len(bytes);
-
+            let len = Char::utf8_len(bytes[0]) as usize;
+            iif![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
             if len > 2 { new.arr[2] = bytes[2]; }
@@ -409,8 +408,8 @@ impl<const CAP: usize> StringNonul<CAP> {
         let mut new = unwrap![ok? Self::new()];
         if !c.is_nul() {
             let bytes = c.to_utf8_bytes();
-            let len = Char::utf8_2bytes_len(bytes);
-
+            let len = Char::utf8_len(bytes[0]) as usize;
+            iif![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
         }
@@ -433,8 +432,8 @@ impl<const CAP: usize> StringNonul<CAP> {
         let mut new = unwrap![ok? Self::new()];
         if !c.is_nul() {
             let bytes = c.to_utf8_bytes();
-            let len = Char::utf8_3bytes_len(bytes);
-
+            let len = Char::utf8_len(bytes[0]) as usize;
+            iif![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
             if len > 2 { new.arr[2] = bytes[2]; }